Create preservation flag for servers.sh.

+++ servers.sh
@@ -4,11 +4,19 @@
 # You can customize the nr of servers to run and the starting 
 # range of the TCP/IP ports. ps: 48 is nice since we have 16 levels.
 
-# Kill all running servers
-killall -9 smashbattle
+# usage: bash servers.sh
+#    or: bash servers.sh preserve
+#
+# preservation flag will make sure the servers that are already running
+# will keep running
 
-# Wait for it ...
-sleep 2
+if [ "$1" != "preserve" ]; then
+	# Kill all running servers
+	killall -9 smashbattle
+
+	# Wait for it ...
+	sleep 2
+fi
 
 # Number of servers we want to start
 servercount=32
@@ -44,13 +52,19 @@ levelcount=0
 
 while [  $count -lt $servercount ]; do
 
-        # Launch a server
-        printf "Starting Tweak Battle Server #$count \n"
-        (daemon -- smashbattle -s "${levels[$levelcount]}" $port "${levels[$levelcount]}" > /var/log/smashbattle/server-${count}) &
-        sleep 1
+		# Launch a server
+		printf "Starting Tweak Battle Server #$count \n"
 
-        # Increment port, loopcount and count
-        port=$((port+1))
-        count=$((count+1))
-	levelcount=$(((levelcount+1) % 16))
+		if [ "$1" != "preserve" ]; then
+				(daemon -- smashbattle -s "${levels[$levelcount]}" $port "${levels[$levelcount]}" > /var/log/smashbattle/server-${count}) &
+		else
+				(daemon -- smashbattle -s "${levels[$levelcount]}" $port "${levels[$levelcount]}" >> /var/log/smashbattle/server-${count}) &
+		fi
+
+		sleep 1
+
+		# Increment port, loopcount and count
+		port=$((port+1))
+		count=$((count+1))
+		levelcount=$(((levelcount+1) % 16))
 done
